+++ lib/flutter_shape_image.dart
@@ -1,3 +1,4 @@
 library flutter_shape_image;
 
 export 'src/circle_image.dart';
+export 'src/diamond_image.dart';