+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ## [0.0.1] - 05 Sept 2018
 
-* Create Circle Image
+* Add Circle Image
+
+## [0.0.2] - 05 Sept 2018
+
+* Add Diamond Image